Move Kademlia protocol functions to protocol module
